+++ p_doors.c
@@ -191,7 +191,7 @@ void T_VerticalDoor (MEMREF memref)
 
 int16_t
 EV_DoLockedDoor
-( int16_t linetag,
+( uint8_t linetag,
 	int16_t linespecial,
   vldoor_e	type,
   MEMREF thingRef )
@@ -250,7 +250,7 @@ EV_DoLockedDoor
 
 int16_t
 EV_DoDoor
-( int16_t linetag,
+( uint8_t linetag,
   vldoor_e	type )
 {
     int16_t		secnum,rtn;
